Fix #124
Allows passengers time to enter or exit elevators if their indicator
lights change.

This fixes a common issue where people change the elevator indicator
lights in the on("idle") callback at the same time as they goToFloor, so
passengers waiting for the elevator do not have time to board.

The idea here is that the programmer does not control the elevator doors
directly. The lift won't move if the doors are open and the doors won't
close if they are obstructed, so passengers can get on and off the lifts
if the indicator lights change.

The whole model is kinda callback hell, but I wanted to make a small PR
and didn't change that.